Function that compares hyper-parameter tuning
If you run the last section after doing the hyper-parameter tuning, it compares the model with the best parameters to a base model. It finds that the base model does better with the test data than the tuned model which means we are over-fitting the training data.  This is proven by doing the same test but using the training data which produces about the same error.  Currently stuck on trying to figure out what to do from here.