Fixed Overfitting Issue
Can now get random forest down to RMSE of 41 compared to 43 without hyper parameter tuning.  Tried linear model and only got 44. So the question now is: is 41 good enough? Working on doing gridsearchCV around the parameter randomly found but have a feeling that won't give us much of a boost.  Thus, I propose we plot the efficiency of tuning our trees and talk about that.  Clean it up and make some visuals and then publish it saying we would like to explore neural nets next but this is where were at right now.